requirements.txt
pyyaml
easydict
tqdm
numpy
matplotlib
pillow